requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.0.12
chromedriver-py==109.0.5414.25
discord==2.1.0
discord-py-slash-command==4.2.1
discord.py==2.0.1
dnspython==2.2.1
exceptiongroup==1.0.4
frozenlist==1.3.1
graphviz==0.20.1
h11==0.14.0
idna==3.4
multidict==6.0.2
outcome==1.2.0
ply==3.11
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
yarl==1.8.1